performance_agent/requirements.txt
httpx==0.28.1
fastapi==0.116.0
pydantic==2.11.7
uvicorn==0.35.0
google-generativeai==0.8.5